make parser and move a lot of functions around into CLIController
