+++ maven/initial/src/main/java/uml/GUIController.java
@@ -1,4 +1,9 @@
 package uml;
 
 public class GUIController {
+
+    public static void main(String[] args) {
+        
+    }
+
 }
